monitor/server/pyproject.toml
[project]
name = "server-results-provider"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "matplotlib>=3.10.7",
    "pika>=1.3.2",
    "shared-lib",
]

[tool.uv.sources]
shared-lib = { path = "../shared_lib", editable = true }
